+++ frontend/map-corridors/src/App.tsx
@@ -1343,6 +1343,7 @@ function App() {
             activePhotoMarkerId={activePhotoMarkerId}
             onActivePhotoMarkerChange={setActivePhotoMarkerId}
             onPhotoPreview={handleOpenPhotoPreview}
+            onCompareVariants={handleCompareVariants}
             provisionalPlacement={provisionalPlacement}
             onProvisionalDrag={handleProvisionalDrag}
             onProvisionalCommit={handleProvisionalCommit}